requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-taggit==6.1.0
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
oauthlib==3.2.2
psycopg==3.2.3
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
